Move all Configuration Component to EmberCore
