+++ EmberCore/KernelServices/UI/View/Configuration/Components/CheckComponent.xaml
@@ -1,9 +1,9 @@
-﻿<CheckBox x:Class="EmberCore.KernelServices.UI.ViewModel.Configuration.Component.CheckComponent"
+﻿<CheckBox x:Class="EmberCore.KernelServices.UI.View.Configuration.Components.CheckComponent"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:converts="clr-namespace:EmberWpfCore.Components.Configuration.View.Component.Converts"
+             xmlns:converts="clr-namespace:EmberCore.KernelServices.UI.View.Configuration.Converts"
              mc:Ignorable="d" 
              d:DesignHeight="450" d:DesignWidth="800">
     <CheckBox.Resources>
