+++ EmberCore/KernelServices/UI/View/Configuration/Components/ComboComponent.xaml
@@ -1,4 +1,4 @@
-﻿<ComboBox x:Class="EmberCore.KernelServices.UI.ViewModel.Configuration.Component.ComboComponent"
+﻿<ComboBox x:Class="EmberCore.KernelServices.UI.View.Configuration.Components.ComboComponent"
             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
