+++ EmberCore/KernelServices/UI/View/Configuration/Components/IntegerComponent.xaml
@@ -1,12 +1,15 @@
-﻿<TextBox x:Class="EmberCore.KernelServices.UI.ViewModel.Configuration.Component.StringComponent"
+﻿<TextBox x:Class="EmberCore.KernelServices.UI.View.Configuration.Components.IntegerComponent"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:converts="clr-namespace:EmberWpfCore.Components.Configuration.View.Component.Converts"
+             xmlns:converts="clr-namespace:EmberCore.KernelServices.UI.View.Configuration.Converts"
              mc:Ignorable="d" 
              d:DesignHeight="450" d:DesignWidth="800">
+    <TextBox.Resources>
+        <converts:IntValueConverter x:Key="Converter"/>
+    </TextBox.Resources>
     <TextBox.Text>
-        <Binding Path="Value"></Binding>
+        <Binding Path="Value" Converter="{StaticResource Converter}"></Binding>
     </TextBox.Text>
 </TextBox>
