+++ EmberCore/KernelServices/UI/View/Configuration/Components/StringComponent.xaml
@@ -1,15 +1,12 @@
-﻿<TextBox x:Class="EmberCore.KernelServices.UI.ViewModel.Configuration.Component.DoubleComponent"
+﻿<TextBox x:Class="EmberCore.KernelServices.UI.View.Configuration.Components.StringComponent"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:converts="clr-namespace:EmberWpfCore.Components.Configuration.View.Component.Converts"
+             xmlns:converts="clr-namespace:EmberCore.KernelServices.UI.View.Configuration.Converts"
              mc:Ignorable="d" 
              d:DesignHeight="450" d:DesignWidth="800">
-    <TextBox.Resources>
-        <converts:DoubleValueConverter x:Key="Converter"/>
-    </TextBox.Resources>
     <TextBox.Text>
-        <Binding Path="Value" Converter="{StaticResource Converter}" Mode="TwoWay"></Binding>
+        <Binding Path="Value"></Binding>
     </TextBox.Text>
 </TextBox>
